# docs/swagger.yaml
basePath: /v1
definitions:
...
  dtos.CreateProductRequest:
    properties:
      category:
        type: string
      description:
        type: string
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
    required:
    - category
    - description
    - name
    - price
    - stock
    type: object
...
  dtos.ErrorResponse:
    properties:
      error:
        example: Validation failed
        type: string
    type: object
...
  dtos.PatchProductRequest:
    properties:
      category:
        type: string
      description:
        type: string
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
    type: object
  dtos.ProductDetail:
    properties:
      category:
        type: string
      created_at:
        example: "2024-12-26T01:59:44.840049+01:00"
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
      updated_at:
        example: "2024-12-26T01:59:44.840049+01:00"
        type: string
    type: object
  dtos.ProductListResponse:
    properties:
      page:
        example: 2
        type: integer
      page_size:
        example: 10
        type: integer
      products:
        items:
          $ref: '#/definitions/dtos.ProductDetail'
        type: array
      total_count:
        example: 100
        type: integer
      total_pages:
        example: 10
        type: integer
    type: object
...
  models.Product:
    properties:
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a sample e-commerce API documentation.
  title: E-Commerce API
  version: "1.0"
paths:
...
  /products:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of products with the ability to specify
        page and page size.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of products per page
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved the paginated list of products
          schema:
            $ref: '#/definitions/dtos.ProductListResponse'
        "400":
          description: Invalid page number or pageSize
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: Retrieve a paginated list of products
      tags:
      - Product
    post:
      consumes:
      - application/json
      description: Allows an admin to create a new product
      parameters:
      - description: Product information
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Product created successfully
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "401":
          description: Unauthenticated, login is required
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "403":
          description: Unauthorized access, only admins can create products
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new product
      tags:
      - Product
  /products/{id}:
    delete:
      description: Allows an admin to delete a product by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Product deleted successfully
        "400":
          description: Invalid product ID
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "401":
          description: Unauthenticated, login is required
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "403":
          description: Unauthorized access, only admins can delete products
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a product
      tags:
      - Product
    get:
      consumes:
      - application/json
      description: Retrieve a product by its unique ID.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.ProductDetail'
        "400":
          description: Invalid product ID
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: Retrieve a product by ID
      tags:
      - Product
    patch:
      consumes:
      - application/json
      description: Allows an admin to update specific fields of an existing product
        by providing the product ID and the updated data.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product data to patch
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dtos.PatchProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated product details
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Invalid product ID or request payload
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "401":
          description: Unauthenticated, login is required
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "403":
          description: Unauthorized access, only admins can patch products
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Partially update an existing product
      tags:
      - Product
    put:
      consumes:
      - application/json
      description: Allows an admin to fully update all fields of an existing product
        by providing the product ID and new data.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product data to update
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated product details
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Invalid product ID or request payload
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "401":
          description: Unauthenticated, login is required
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "403":
          description: Unauthorized access, only admins can update products
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Fully update an existing product
      tags:
      - Product
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"